Converted code to MVC Architechture
