+++ package.json
@@ -2,7 +2,7 @@
   "name": "real-review",
   "version": "1.0.0",
   "description": "This is an internship project",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
